refactor: 패키지 순환참조 끊기
