+++ src/main/java/com/likelion/mooding/feedback/application/FeedbackChatCompletionService.java
@@ -1,7 +1,7 @@
 package com.likelion.mooding.feedback.application;
 
 import com.likelion.mooding.feedback.application.dto.FeedbackCreateResponse;
-import com.likelion.mooding.feedback.presentation.dto.FeedbackCreateRequest;
+import com.likelion.mooding.feedback.application.dto.FeedbackCreateRequest;
 import reactor.core.publisher.Mono;
 
 public interface FeedbackChatCompletionService {
