+++ src/main/java/com/likelion/mooding/feedback/application/FeedbackService.java
@@ -6,9 +6,9 @@ import com.likelion.mooding.feedback.domain.Feedback;
 import com.likelion.mooding.feedback.domain.FeedbackRepository;
 import com.likelion.mooding.feedback.domain.FeedbackStatus;
 import com.likelion.mooding.feedback.exception.FeedbackAuthException;
-import com.likelion.mooding.feedback.presentation.dto.FeedbackCreateRequest;
-import com.likelion.mooding.feedback.presentation.dto.FeedbackResultResponse;
-import com.likelion.mooding.feedback.presentation.dto.FeedbackStatusResponse;
+import com.likelion.mooding.feedback.application.dto.FeedbackCreateRequest;
+import com.likelion.mooding.feedback.application.dto.FeedbackResultResponse;
+import com.likelion.mooding.feedback.application.dto.FeedbackStatusResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import reactor.core.publisher.Mono;
