+++ src/main/java/com/likelion/mooding/feedback/presentation/FeedbackController.java
@@ -3,9 +3,9 @@ package com.likelion.mooding.feedback.presentation;
 import com.likelion.mooding.auth.annotation.Auth;
 import com.likelion.mooding.auth.presentation.dto.Guest;
 import com.likelion.mooding.feedback.application.FeedbackService;
-import com.likelion.mooding.feedback.presentation.dto.FeedbackCreateRequest;
-import com.likelion.mooding.feedback.presentation.dto.FeedbackResultResponse;
-import com.likelion.mooding.feedback.presentation.dto.FeedbackStatusResponse;
+import com.likelion.mooding.feedback.application.dto.FeedbackCreateRequest;
+import com.likelion.mooding.feedback.application.dto.FeedbackResultResponse;
+import com.likelion.mooding.feedback.application.dto.FeedbackStatusResponse;
 import java.net.URI;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
