+++ src/test/java/com/likelion/mooding/feedback/application/FeedbackServiceTest.java
@@ -11,8 +11,8 @@ import com.likelion.mooding.feedback.domain.Feedback;
 import com.likelion.mooding.feedback.domain.FeedbackRepository;
 import com.likelion.mooding.feedback.domain.FeedbackStatus;
 import com.likelion.mooding.feedback.exception.FeedbackAuthException;
-import com.likelion.mooding.feedback.presentation.dto.FeedbackCreateRequest;
-import com.likelion.mooding.feedback.presentation.dto.FeedbackStatusResponse;
+import com.likelion.mooding.feedback.application.dto.FeedbackCreateRequest;
+import com.likelion.mooding.feedback.application.dto.FeedbackStatusResponse;
 import java.time.Duration;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
